Make r10edocker version more easily set (#10)
* Make r10edocker version more easily set

We make the version of `r10edocker` more easily overridable during
build, which reduces the amount of error-prone manual work to allow
better automation.

* Un-dockerignore the .git directory

We need to get the .git directory copied to Docker builder for the version info.

* Update release.yml

With the latest changes, released app version should always be identical
to the tagged version. Therefore, we update the version check failure
message as if it's an assert.