Fetch all history for all tags (#15)
* Fetch all history for all tags

In `canary.yml`, fetch all commit history for tags. Unfortunately,
there's no easy way to make canary artifacts display a fixed "canary"
version string, due to the way we set `VERSION` value in Makefile. But
this should be OK as long as the canary builds are reproducible.

* Add debugging info in canary.yml

Show version string of built artifact.

* Fetch all commit history and tags in sanity.yml

Because we use Git tag info to set app version, we need to fetch this
info.