Use newer dockerhub nixos/nix image as base image
Use the current latest of nixos/nix container image as the base image
for the Dockerfile template.

I heard that the r10e-docker build had an issue on macOS/M3 when the
older base image was used, and switching to this latest version resolved
the issue.

Thank @wmagda for reporting the issue and finding the fix.