Provide config option to include cacert
We provide a new `include_ca_bundle` option in r10edocker's
configuration file to determine if a root CA bundle will be included in
the generated Docker container image or not. The CA bundle is usually
necessary if the contained application uses TLS.

We also update documentation and the sample "config.json" file to show
case this new option.